requirements.txt
flask
argon2-cffi
mysql-connector-python
